Update dynamic paraetrization to use proper pytest api
This removes a new deprecation warning and makes the test
setup align with pytest documentation on dynamic parametrization.
