+++ jwst/tests_nightly/general/associations/conftest.py
@@ -1,6 +1,9 @@
 """Pytest configurations"""
 import pytest
 
+from jwst.tests_nightly.general.associations.sdp_pools_source import SDPPoolsSource
+
+
 # Add option to specify a single pool name
 def pytest_addoption(parser):
     parser.addoption(
@@ -12,3 +15,19 @@ def pytest_addoption(parser):
 @pytest.fixture
 def sdp_pool(request):
     return request.config.getoption('--sdp-pool')
+
+
+def pytest_generate_tests(metafunc):
+    """Prefetch and parametrize a set of test pools"""
+    if 'pool_path' in metafunc.fixturenames:
+        SDPPoolsSource.inputs_root = metafunc.config.getini('inputs_root')[0]
+        SDPPoolsSource.results_root = metafunc.config.getini('results_root')[0]
+
+        pools = SDPPoolsSource()
+
+        try:
+            pool_paths = pools.pool_paths
+        except Exception:
+            pool_paths = []
+
+        metafunc.parametrize('pool_path', pool_paths)
